worker_node/requirements.txt
grpcio
grpcio-tools
ffmpeg-python